requirements.txt
# Core exchange
ccxt>=4.3.0

# Data & numerics
pandas>=2.0.0
numpy>=1.26.0

# Config
PyYAML>=6.0

# Telegram alerts (optional — bot works without this)
requests>=2.31.0

# Dev / testing
pytest>=8.0.0
pytest-mock>=3.12.0
